Make the create-extractor-pack.ps1 script more reliable.

+++ create-extractor-pack.ps1
@@ -3,7 +3,9 @@ cargo build --release
 cargo run --release -p ql-generator
 codeql query format -i ql\src\codeql_ql\ast\internal\TreeSitter.qll
 
-rm -Recurse -Force extractor-pack
+if (Test-Path -Path extractor-pack) {
+	rm -Recurse -Force extractor-pack
+}
 mkdir extractor-pack | Out-Null
 cp codeql-extractor.yml, ql\src\ql.dbscheme, ql\src\ql.dbscheme.stats extractor-pack
 cp -Recurse tools extractor-pack
